pyproject.toml
[tool.poetry]
name = "anoscout"
version = "0.1.0"
description = ""
authors = ["Julian Rakuschek <julian.rakuschek@tugraz.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
flask = "^3.1.0"
pymongo = "^4.11.1"
redis = "^5.2.1"
flask-cors = "^5.0.0"
numpy = "<2"
aeon = "^1.0.0"
scikit-learn = "<1.6.0"
tslearn = "^0.6.3"
scipy = "<1.15"
fastdtw = "^0.3.4"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
h5py = "^3.13.0"
pyod = "^2.0.3"
psutil = "^7.0.0"
timeeval-gutentag = "^1.4.2"
keras = "^3.8.0"
tensorflow = "^2.18.0"
stumpy = "^1.13.0"
dtaidistance = "^2.3.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
